+++ src/main/java/net/browsertime/tool/BasicAuth.java
@@ -1,0 +1,7 @@
+package net.browsertime.tool;
+
+public class BasicAuth {
+  public String domain;
+  public String username;
+  public String password;
+}